Refactor ChangePhoneModal
